separate order from shopping_bag
