+++ models/shopping_bag.py
@@ -29,26 +29,3 @@ class BagItem(Base):
 
     def __repr__(self):
         return f"<BagItem - (product_id={self.product_id}, quantity={self.quantity})>"
-
-
-
-class OrderStatus(Enum):
-    PREPARING = "preparing"
-    ACCEPTED = "accepted"
-    SENT = "sent"
-    CANCELED = "canceled"
-    REJECTED = "rejected"
-
-class Order(Base):
-    __tablename__ = "orders"
-
-    id = Column(Integer, primary_key=True, index=True)
-    user_profile_id = Column(Integer, ForeignKey("user_profiles.id", ondelete="CASCADE"), nullable=False)
-    shopping_bag_id = Column(Integer, ForeignKey("shopping_bags.id", ondelete="CASCADE"), nullable=False)
-    created = Column(DateTime, default=datetime.utcnow)
-    order_number = Column(String, unique=True, nullable=False)
-    is_paid = Column(Column(Boolean, default=False))
-    status = Column(Enum(OrderStatus), nullable=False)
-
-    user_profile = relationship("UserProfile", back_populates="orders")
-    shopping_bag = relationship("ShoppingBag", back_populates="order")
